Multiprocessing!
Start a function in the background that continuously attempts to
read a byte from the serial port. Any bytes it successfully reads
get put in a queue for retrieval. In this way we can read the full
response from the halfy without needing to know what it will be
ahead of time.

Todo: Literally everything...