Config stays in the dictionary. Inputs, Outputs are dictionaries.
* The config values stay in the config{} dictionary. This makes for
less passing of variables around and makes references to config
variables explicit. As before, the inputs, outputs, and levels
have to be made integers.

* The inputs and outputs config keys now contain dictionaries of
the format {[port number]:[label]}. This allows for gaps in the
ports used. Instead of iterating over the ports with range(), we
can directly iterate over the dictionaries and we can ask if a
port is in the inputs or outputs objects. We can also use the
label values when presenting our results.