One fewer process and stability...?
Dropped the use of multiprocessing for reading from the halfy.
Instead, we just use time.sleep() to wait for the response, and we
use halfy.in_waiting to determine how many bytes to read from the
hafly. Also, we do a halfy.reset_input_buffer() to make sure that
there aren't any stray bytes kicking around in the input buffer.
Removed listener() as we don't need it. Simplified init code too.