Smarter command timing... I think...
signaler now gets passed command_type which is the first two chars
of the command string. This is used to determine how many bytes
the halfy should return to us. We wait for that number of bytes to
appear (max 1 second), wait .005 seconds longer (it takes a little
longer for the message to be ready?), and then carry on. This is
instead of the previously-hardcoded .05 second wait and is much
faster in practice.